Basic audio code
I'll actually go through and add comments and stuff, later, but here is what I've been working on.
Currently it can take a poem in, and then it will automatically play music according to the tone of the stanza, the stanza will also automatically be read.
How the reading and music gets called will probably be split into different functions that can be controlled by the UI once we integrate that in.